pyproject.toml
[project]
name = "metwant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyarrow>=22.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
